+++ services/static/src/main/java/soccervs/statics/core/repositories/PosicaoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import soccervs.statics.core.entities.Posicoes;
 
 @Repository
-public interface PosicaoRepository extends JpaRepository<Posicoes, Byte> {
+public interface PosicaoRepository extends JpaRepository<Posicoes, Long> {
 	@Query("SELECT p.id FROM Posicoes p WHERE p = :posicao")
 	Byte findIdByPosicao(@Param("posicao") Posicoes posicao);
