+++ services/static/src/main/java/soccervs/statics/core/repositories/TecnicoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import soccervs.statics.core.entities.Tecnicos;
 
 @Repository
-public interface TecnicoRepository extends JpaRepository<Tecnicos, Integer> {
+public interface TecnicoRepository extends JpaRepository<Tecnicos, Long> {
 	@Query("SELECT t.id FROM Tecnicos t WHERE t = :tecnico")
 	Integer findIdByTecnico(@Param("tecnico") Tecnicos tecnico);
