+++ services/static/src/main/java/soccervs/statics/core/repositories/TituloRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import soccervs.statics.core.entities.Titulos;
 
 @Repository
-public interface TituloRepository extends JpaRepository<Titulos, Integer> {
+public interface TituloRepository extends JpaRepository<Titulos, Long> {
 
 }
